+++ src/Core/AdminConsole/OrganizationFeatures/OrganizationUsers/Interfaces/IRemoveOrganizationUserCommand.cs
@@ -1,4 +1,5 @@
-﻿using Bit.Core.Enums;
+﻿using Bit.Core.Entities;
+using Bit.Core.Enums;
 
 namespace Bit.Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Interfaces;
 
@@ -8,4 +9,6 @@ public interface IRemoveOrganizationUserCommand
 
     Task RemoveUserAsync(Guid organizationId, Guid organizationUserId, EventSystemUser eventSystemUser);
     Task RemoveUserAsync(Guid organizationId, Guid userId);
+    Task<List<Tuple<OrganizationUser, string>>> RemoveUsersAsync(Guid organizationId,
+        IEnumerable<Guid> organizationUserIds, Guid? deletingUserId);
 }
